security: Adjusted API reference.
Me and, therefore, the reference were under the false assumption that
there would be any benefit in hashing the passphrase on the client only.
If this is done the hash in effect becomes the passphrase, and if the
server doesn't rehash it we are in effect storing the password in plain.
That is, this mecanhism would be **less** secure than the standard
mechanism used in most software. There's a reason they say to not come
up with your own security.

To add insult to injury the API reference mentioned that the passphrase
would not be sent to the server with capable clients, it would only use
the passphrase as the "encryption key". There's no such thing in
traditional passphrase hasing functions. There is (or should be) a salt,
but this is just appended to the passphrase and this is encrypted
with the given hash function. That is, the passphrase **would be** sent to the
server, only hashed.

All of this was clarified. There's still a recommendation to do
client-side hashing, but the server always hashes what is sent up with a
salt, even if already hashed.